Check the result of snprintf operations for truncation
Newer versions of GCC warns if truncation of string operations is
possible and the code does not check for it.  There were some snprintf
operations that were triggering that warning and causing the build to
fail.  The results of the snprintf are now checked to ensure that
truncation did not occur.  This was reported on:

https://sourceforge.net/p/perfmon2/bugs/7/