fix grpid handling for Intel X86 uncore
On SkylakeX the umask grpid field is overloaded to contain two subfield.
The actual grpid and the required grpid (at offset 8). The encoding code
has a bug where it would not use the accessor function get_grpid() to extract
the group id from the field. Given that the grpid is used in statements such
as:
     u = 1 << pe[e->event].umasks[a->idx].grpid;
The code could run the risk of exceeding the max shift for a 16-bit value.
The fix is to use accessor function to extract the grpid.

The patch also adds a validation test to ensure events which would cause
a large grpid are properly encoded.

Signed-off-by: Stephane Eranian <eranian@gmail.com>