Fix AMD64 Fam19h L3 PMU support
The PMU perf_events type was not correctly encoded because the .perf_name field
was not initialized and therefore it defaulted to using the core PMU. The correct
perf_name is "amd_l3". With that in place, the library now picks up the correct
PMU type and associated programming restrictions, e.g., per-cpu mode only and code
such as perf_examples/self should not be allow to succeed at perf_event_open().

Reported-by: Steve Kaufmann <steven.kaufmann@hpe.com>
Signed-off-by: Stephane Eranian <eranian@gmail.com>