do not set certain config bits in pfm_amd64_get_perf_encoding()
By default on Intel X86, the library was setting the EN and INT bits
for each core PMU events. But when encoding for perf_events, these
bits are ignored by the interface and reprogrammed by the kernel.
Similarly, the USR/OS/GUEST/HOST bits are controlled by the perf_event_attr.exclude_*
field not the config field.

Recent versions of the Linux perf tool warn when bits which are ignored
are set in the config field which is useful.

This patch clears all the config bits under the control of the perf_events
interface.

The encoding for raw PMU mode is unchanged.

Signed-off-by: Stephane Eranian <eranian@gmail.com>