do not set certain config bits in pfm_arm_get_perf_encoding()
By default (raw encoding) on ARM, the library was setting the PL1, USR, HYP control bits
in the config in the encoded value.  With Linux perf_events, these bits are under the control
of the kernel. Any of these bits set by the user is overridden by the kernel based on the settings
of the perf_event_attr.exclude_* fields. Recent versions of the perf tool started checking that the
config field is not setting bits which are ignored by the kernel. To avoid the perf tool warning,
this patch removes the setting of these bits when encoding for Linux perf_events.

Signed-off-by: Stephane Eranian <eranian@gmail.com>