add Intel Topdown PERF_METRICS pseudo events for Linux perf_events
This patch adds Icelake and later Topdown events as exported by the
Linux perf_events interface. On Icelake, Topdown toplevel is implemented
by a new MSR called PERF_METRICS that requires special handling in the
Linux perf_events implementation. The 5 toplevel metrics are exported by
the kernel via pseudo evnet published in sysfs. This patch makes these
events available via the perf:: PMU in libpfm4 if they are available
on the underlying hardware.

The following topdown events are supported by this patch:
slots
topdown-retiring
topdown-fe-bound
topdown-be-bound
topdown-bad-spec

The patch makes these events available via libpfm4 ony if the kernel exports
them, i.e., the hardware supports them.

In order to use the topdown toplevel events on processors with PERF_METRICS,
the events must be passed to perf_events as a single group with slots as the
leader:

$ perf stat --pfm-events '{slots,topdown-retiring,topdown-fe-bound,topdown-be-bound,topdown-bad-spec}' ....

This requirement comes from the kernel and not libpfm4 or the perf tool.

Signed-off-by: Stephane Eranian <eranian@gmail.com>