fix allocation logic in perf_table_alloc_event()
Because of the addition of optional events in commit

a33aa1a9ee70 add Intel Topdown PERF_METRICS pseudo events for Linux perf_events

There was a problem with the static events in the logic in
perf_table_alloc_event(). The code could end up trying to free
perf_pe when it was pointing to the static event tables. Similarly,
it would try to realloc an array that was never allocated.

This patch fixes the problem and cleans up a few minor problems.

Signed-off-by: Stephane Eranian <eranian@gmail.com>