fix perf_events raw encoding handling of event strings
The pfm_perf_raw_match_event() was relying on sscanf to convert the
raw hexadecimal string into integer.
The problem with sscanf is that it stops at the first non hex character
but does not tell you where. This caused problems with wrongly named
events starting with 'r'. For instance, on AMD Zen3, the event
retired_uops does not exist, yet libpfm4 would encode at 0xe.
That is because the event starts with 'r' and therefore if libpfm4 fails
to find a match in the AMD Zen3 PMU, it will use the raw perf_event syntax
and the seocnd letter is 'e'. Then sscanf stops at the 't' and does not
report any error.

Fix the problem by using strtoull() instead. It returns an end pointer.
In case the string contains non hex, that pointer will point to it.

With this fix in place:
$ examples check_events retired_uops
fails on AMD Zen3.

Signed-off-by: Stephane Eranian <eranian@gmail.com>