avoid GCC-12 use-after-free warnings
gcc-12 seems to complain about bogus use-after-free situations in the
libpfm4 code:

    p = realloc(q, ...)
    if (!p)
       return NULL

    s = p + (q - z)

It complains because of the use of q after realloc in this case.
Yet  q - z is just pointer artihmetic and is not dereferencing any
memory through the pointer q which may have been freed by realloc.

Fix is to pre-computer the delta before realloc to avoid using the
pointer after the call.

Reported-by: Vitaly Chikunov <vt@altlinux.org>
Signed-off-by: Stephane Eranian <eranian@gmail.com>