maintain list of active PMUs
Given that the list of PMUs supported per architecture keeps growing,
it is becoming expensive to iterate over each PMU looking for a
match. The macro pfmlib_for_each_pmu() is iterating over active and
inactive PMUs looking for active ones. Given that the number of inactive
PMUs is always larger than the number of active, this was expensive.

Fix this by creating a list of active PMUs and adding a new macro
pfmlib_for_each_active_pmu(). We use a oubly linked list to allow
further optimizations.

As an example on a X86 build, the new iterator allowed a 10x reduction
in iterations inside pfmlib_parse_event() for a core PMU event.

When LIBPFM_ENCODE_INACTIVE is set to 1, then all PMUs supported by
the architecture are put on the active list even when they are not
detected. This simplifies the parsing loop.

Signed-off-by: Stephane Eranian <eranian@gmail.com>