fix active PMU list handling for LIBPFM_ENCODE_INACTIVE=1
There was an issue introduced by:
 363825f72afd ("maintain list of active PMUs")

Where if a PMU is not detected because not exported by OS, it would
not be put on the active list when LIBPFM_ENCODE_INACTIVE=1 causing
tests/validate on some uncore PMU events.

Fix this by correctly handling the case where a PMU is not exported
by the OS.

Also check that a PMU is actually active in pfmlib_terminate() and
pfm_get_pmu_by_type() again to handle LIBPFM_ENCODE_INACTIVE=1.

Signed-off-by: Stephane Eranian <eranian@gmail.com>