Revert "optimize active PMU list further"
This reverts commit b009b1263098eec925bc2dba1760c70d8a46d4b8.

Because it makes it necessary to have a lock as the head of the
list may be changing at each encoding and that would cause issues
with multiple parallel calls of the encoding entry points.

This optimization will be redone with a thread local variiable
once we modify libpfm4 to depend on libpthread.

Signed-off-by: Stephane Eranian <eranian@gmail.com>