fix active list ordering issue
In commit
	363825f72afd ("maintain list of active PMUs")

We introduce an active list of PMUs to speed up lookups of events.
However, there was a bug introduced by this commit which caused
wrong encodings of certain events. For instance, on Intel x86,
the event unhalted_reference_cycles was encoded using an architected
event code instead of the model specific event code which used a fixed
counter. This was due to the fact that the ordering of the PMU models
in pfmlib_pmus[] was inverted by the way we built the active list. The
order in the table matters for lookups. the list must maintain the same
order. This patch fixes the problem by rewriting the linked list code
to support appending to the tail of the list (instead of the head).
That way the order in the table is maintained.

The patch introduces the notion of a link list node supporting a
double linked list data structure with basic accessor functions.

Signed-off-by: Stephane Eranian <eranian@gmail.com>