add CONFIG_PFMLIB_NOTRACEPOINT to speedup libpfm4 initialization
When pfmlib_initialize() is run as root and if the debugs is mounted,
then the library parses all the tracepoints to add them to the perf PMU.
Depending on the number of tracepoints, this can take a significant amount
of time even though this may not be needed if no tracepoints is passed.
In order to speedup pfm_initialize(), the patch adds a compile-time option
to disable support for tracepoint in the perf PMU.

To deactivate tracepoint support: make CONFIG_PFMLIB_NOTRACEPOINT=y

The default build is unchanged with tracepoints enabled. This patch
adds an opt-out option.

Signed-off-by: Stephane Eranian <eranian@gmail.com>