Add support_no_mods support to perf_events encoding
This patch adds handling of the support_no_mods attribute info to
perf_events encoding routine. If the flag is set for any umasks
then it is applied to all. It is expected that events with such
conditions prohibit umasks combinations. Only the perf_events
pinned attribute is maintained because it is a modifier of the
perf_events subsystem with no specific implication in the hardware.

Signed-off-by: Stephane Eranian <eranian@gmail.com>