Add support for Topdown via PERF_METRICS for Intel GraniteRapids
Cut & paste of the Intel SapphireRapids support.

Add the TOPDOWN_M dedicated event to provide the pseudo
encodings necessary to program Topdown L1 and L2 events onto the
PERF_METRICS MSR of Intel GraniteRapids on Linux.

In order to successfully use PERF_METRICS, the kernel imposes
some restrictions (which are not known to libpfm4) which the
user must follow:
      TOPDOWN_M events must be passed to the kernel in a single
      perf_events group (chained fds) AND the TOPDOWN_M.SLOTS event
      must be the first event in that group.

Note that only the SLOTS events (programmed in fixed counter3)
support modifiers such as user vs. kernel, hw_smpl or precise
sampling on perf_events. All other umasks do not support any modifier.
The SLOTS event controls the filtering for all PERF_METRICS pseudo
events.

The encodings provided by libpfm4 for fixed counters are specific
to Linux. When used on non Linux systems, encodings are not
guaranteed to be valid.

Signed-off-by: Stephane Eranian <eranian@gmail.com>